+++ packages/cordova/src/ios/AMBAdBase.swift
@@ -42,7 +42,7 @@ class AMBAdBase: AMBCoreAd {
         self.emit(eventName, ["message": error.localizedDescription])
     }
 
-    func emit(_ eventName: String, _ reward: GADAdReward) {
+    func emit(_ eventName: String, _ reward: AdReward) {
         self.emit(eventName, [
             "reward": [
                 "amount": reward.amount,
@@ -51,7 +51,7 @@ class AMBAdBase: AMBCoreAd {
         ])
     }
 
-    func emit(_ eventName: String, _ adSize: GADAdSize) {
+    func emit(_ eventName: String, _ adSize: AdSize) {
         self.emit(eventName, [
             "size": [
                 "width": adSize.size.width,
@@ -66,7 +66,7 @@ class AMBAdBase: AMBCoreAd {
         plugin.emit(eventName, data: d)
     }
 
-    func emit(_ eventName: String, _ nativeAd: GADNativeAd) {
+    func emit(_ eventName: String, _ nativeAd: NativeAd) {
         plugin.emit(eventName, data: ["adId": nativeAd.hashValue])
     }
 }