+++ packages/cordova/src/ios/AMBRewarded.swift
@@ -1,7 +1,7 @@
 import GoogleMobileAds
 
-class AMBRewarded: AMBAdBase, GADFullScreenContentDelegate {
-    var mAd: GADRewardedAd?
+class AMBRewarded: AMBAdBase, FullScreenContentDelegate {
+    var mAd: RewardedAd?
 
     deinit {
         clear()
@@ -14,46 +14,44 @@ class AMBRewarded: AMBAdBase, GADFullScreenContentDelegate {
     override func load(_ ctx: AMBContext) {
         clear()
 
-        GADRewardedAd.load(withAdUnitID: adUnitId, request: adRequest, completionHandler: { ad, error in
-            if error != nil {
-                self.emit(AMBEvents.adLoadFail, error!)
+        Task { @MainActor in
+            do {
+                let ad = try await RewardedAd.load(with: self.adUnitId, request: self.adRequest)
+                self.mAd = ad
+                ad.fullScreenContentDelegate = self
+                ad.serverSideVerificationOptions = ctx.optServerSideVerificationOptions()
 
-                ctx.reject(error!)
-                return
+                self.emit(AMBEvents.adLoad)
+                ctx.resolve()
+            } catch {
+                self.emit(AMBEvents.adLoadFail, error)
+                ctx.reject(error)
             }
-
-            self.mAd = ad
-            ad?.fullScreenContentDelegate = self
-            ad?.serverSideVerificationOptions = ctx.optGADServerSideVerificationOptions()
-
-            self.emit(AMBEvents.adLoad)
-
-            ctx.resolve()
-        })
+        }
     }
 
     override func show(_ ctx: AMBContext) {
-        mAd?.present(fromRootViewController: plugin.viewController, userDidEarnRewardHandler: {
+        mAd?.present(from: plugin.viewController, userDidEarnRewardHandler: {
             let reward = self.mAd!.adReward
             self.emit(AMBEvents.adReward, reward)
         })
         ctx.resolve()
     }
 
-    func adDidRecordImpression(_ ad: GADFullScreenPresentingAd) {
+    func adDidRecordImpression(_ ad: FullScreenPresentingAd) {
         self.emit(AMBEvents.adImpression)
     }
 
-    func ad(_ ad: GADFullScreenPresentingAd, didFailToPresentFullScreenContentWithError error: Error) {
+    func ad(_ ad: FullScreenPresentingAd, didFailToPresentFullScreenContentWithError error: Error) {
         clear()
         self.emit(AMBEvents.adShowFail, error)
     }
 
-    func adWillPresentFullScreenContent(_ ad: GADFullScreenPresentingAd) {
+    func adWillPresentFullScreenContent(_ ad: FullScreenPresentingAd) {
         self.emit(AMBEvents.adShow)
     }
 
-    func adDidDismissFullScreenContent(_ ad: GADFullScreenPresentingAd) {
+    func adDidDismissFullScreenContent(_ ad: FullScreenPresentingAd) {
         clear()
         self.emit(AMBEvents.adDismiss)
     }